Abstraindo codigo da AI
