+++ Assets/Scripts/IAController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05244cd00bc15194092542f116e4c46f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
